Homework 2 Complete
This homework is based on the first android application homework
from https://developer.android.com/training/basics/firstapp/index.html
It expands only slightly on it by asking the student to import an
image into the drawable folder and reference. The last part of the
homework the student it challenged to use internet resources to
figure out how to add the image.